refractor: separate collection controllers
