+++ src/controllers/account.ts
@@ -1,18 +1,12 @@
 import { ObjectId, UpdateFilter } from "mongodb";
-import { ProfileDataProjection, UserProfileProjection, ProfileSearchResults } from "../db/schema/projections";
-import { USER_STATUS } from "../utils/enums";
-import { MessageObject, MessageUpdate } from "../../@types";
-import { getPostSignedURL, removeAsset, removeDirectory, removeProfileImage } from "../utils/library/cloudflare";
 import db from "../db/client";
+import { ProfileDataProjection, ProfileSearchResults, UserProfileProjection } from "../db/schema/projections";
+import { USER_STATUS } from "../utils/enums";
+import { MessageObject } from "../../@types";
+import { createNewChat } from "./chat";
 
-//TODO: Separate the functions based on collections
-const chatsCollection = db.collection("chats"),
-  usersCollection = db.collection("users");
-/*
- !    API Functions
-        -> Functions that are used to provide the initial API response
- ?      User Profile
- */
+const usersCollection = db.collection("users");
+
 async function getProfileById<T extends string | string[]>(id: T, myOwnProfile = false) {
   let data: any = null;
   if (typeof id === "string") {
@@ -51,7 +45,6 @@ async function getConnectionData(userId: string, connectionId: string) {
   return data?.connectionData;
 }
 
-//?   Login
 async function verifyUser(usernameOrEmail: string, password: string) {
   const user = await usersCollection.findOne({
     $or: [{ username: usernameOrEmail }, { email: usernameOrEmail }],
@@ -68,7 +61,7 @@ async function verifyUser(usernameOrEmail: string, password: string) {
 }
 
 //     Connections Data for initial request
-async function connectionsData(userId: string) {
+async function getConnectionsData(userId: string) {
   const data = await usersCollection.findOne({ _id: new ObjectId(userId) }, { projection: { _id: 0, connections: 1 } });
   if (!data) return { connections: {}, chats: [], chatIds: [], hasData: false };
   const connections: {
@@ -94,94 +87,14 @@ async function connectionsData(userId: string) {
     )
     .toArray();
 
-  const seenByPropertyMapped: {
-    [userId: string]: boolean;
-  } = {};
   contactsArray.forEach((contact) => {
     //*  Adding profileData for each connection
-    seenByPropertyMapped[connections[contact.id].chat_id] = contact.unseen_messages_count === 0;
     delete contact.unseen_messages_count;
     Object.assign(connections[contact.id], contact);
   });
 
-  const chatIds = Object.values(connections).map((chat) => new ObjectId((chat as { chat_id: string }).chat_id));
 
-  const chatsArray = await getChat(chatIds);
-  chatsArray.forEach((chat, index) => {
-    chatsArray[index].seenByConnection = seenByPropertyMapped[chat.chat_id];
-  });
-
-  return {
-    connections,
-    chats: chatsArray,
-    chatIds: chatIds.map((objectId) => objectId.toString()),
-    hasData: contactIds.length > 0 && chatsArray.length > 0,
-  };
-}
-
-async function getChat(chatIds: ObjectId[] = [], initialMessagesCount = 20) {
-  return await chatsCollection //*  Providing chat data with last 20 messages
-    .find(
-      { _id: { $in: chatIds } },
-      {
-        projection: {
-          _id: 0,
-          chat_id: "$_id",
-          participants: 1,
-          last_updated: 1,
-          created_at: 1,
-          authors_typing: 1,
-          messages: { $slice: ["$messages", -1 * initialMessagesCount] },
-          created_by: 1,
-        },
-      },
-    )
-    .toArray();
-}
-
-/**
- *
- * @param chatId Chat Id in string
- * @param count Number of messages user has
- * @param messagesCount Number of messages user is requesting
- */
-async function getMessages(chatId: string, count = 20, messagesCount = 50) {
-  const data = await chatsCollection
-    .aggregate([
-      {
-        $match: {
-          _id: new ObjectId(chatId),
-        },
-      },
-      {
-        $addFields: {
-          messageCount: { $size: "$messages" },
-        },
-      },
-      {
-        $addFields: {
-          offset: {
-            $subtract: ["$messageCount", count],
-          },
-          pageSize: {
-            $min: [{ $subtract: ["$messageCount", count] }, messagesCount],
-          },
-        },
-      },
-      {
-        $project: {
-          _id: 0,
-          messages: {
-            $slice: ["$messages", { $subtract: ["$offset", "$pageSize"] }, "$pageSize"],
-          },
-          hasMore: { $gt: ["$offset", "$pageSize"] },
-        },
-      },
-    ])
-    .toArray();
-  if (!data?.[0]) return {};
-
-  return data[0];
+  return connections;
 }
 
 async function isUsernameAvailable(user_provided_username: string) {
@@ -301,23 +214,6 @@ async function createNewAccount({
   return newUser;
 }
 
-async function createNewChat(creatorId: string, messageObject: MessageObject | null = null) {
-  if (messageObject) {
-    messageObject.timestamp = new Date();
-    messageObject.id = new ObjectId();
-    messageObject.seenByRecipients = [];
-  }
-  const newChat = await chatsCollection.insertOne({
-    authors_typing: [],
-    created_at: new Date(),
-    last_updated: new Date(),
-    messages: messageObject ? [messageObject] : [],
-    participants: [new ObjectId(creatorId)],
-    created_by: new ObjectId(creatorId),
-  });
-  return newChat;
-}
-
 async function setProfilePictureUrl(user_id: string, url: string) {
   return usersCollection.updateOne({ _id: new ObjectId(user_id) }, { $set: { "avatar.key": url } });
 }
@@ -364,30 +260,6 @@ async function addConnection(fromContactId: string, toContactId: string, message
   };
 }
 
-async function addMessage(chat_id: string, messageObject: MessageObject) {
-  const id = new ObjectId();
-  await chatsCollection.updateOne(
-    { _id: new ObjectId(chat_id) },
-    {
-      $push: { messages: { ...messageObject, id } },
-      $set: { last_updated: messageObject.timestamp },
-    },
-  );
-  return id.toString();
-}
-async function updateSeenMessages(chat_id: string, seenById: string, messageId: string) {
-  return await chatsCollection.updateOne(
-    {
-      _id: new ObjectId(chat_id),
-      "messages.id": new ObjectId(messageId),
-    },
-    {
-      $addToSet: {
-        "messages.$.seenByRecipients": new ObjectId(seenById),
-      },
-    },
-  );
-}
 async function updateUnseenMsgCount(senderId: string, receiverId: string, INCREASE = true) {
   const update = {
     [INCREASE ? "$inc" : "$set"]: {
@@ -428,20 +300,8 @@ async function updateProfile(
   );
 }
 
-async function clearChat(chat_id: string, from: string, to: string) {
-  removeDirectory(`chat_${chat_id}`);
-  return await Promise.all([
-    chatsCollection.updateOne(
-      {
-        _id: new ObjectId(chat_id),
-      },
-      {
-        $set: {
-          messages: [],
-        },
-      },
-    ),
-    usersCollection.updateOne(
+async function clearChatMessageCount(from: string, to: string) {
+  return usersCollection.updateOne(
       {
         _id: new ObjectId(to),
       },
@@ -450,12 +310,10 @@ async function clearChat(chat_id: string, from: string, to: string) {
           [`connections.${from}.unseen_messages_count`]: 0,
         },
       },
-    ),
-  ]);
+    );
 }
 
-async function deleteChat(chatId: string, fromId: string, connectionId: string, toBlock = false) {
-  removeDirectory(`chat_${chatId}`);
+async function deleteChatConnections(fromId: string, connectionId: string, toBlock = false) {
   const updateQuerySender: {
     $unset: object;
     $push?: object;
@@ -494,11 +352,6 @@ async function deleteChat(chatId: string, fromId: string, connectionId: string, 
       },
     ]),
   ]);
-  if (chatId) {
-    await chatsCollection.deleteOne({
-      _id: new ObjectId(chatId),
-    });
-  }
   return;
 }
 
@@ -517,19 +370,6 @@ async function updateStatus(userId: string, { code, update_type }: { code: keyof
       _id: new ObjectId(userId),
     },
     update,
-  );
-}
-
-async function acceptMessageRequest(chatId: string, accepterId: string) {
-  return chatsCollection.updateOne(
-    {
-      _id: new ObjectId(chatId),
-    },
-    {
-      $push: {
-        participants: new ObjectId(accepterId),
-      },
-    },
   );
 }
 
@@ -552,58 +392,6 @@ async function isUserRestricted(restrictId: string, userId: string) {
   ]);
 
   return result && (result[0]?._id || result[1]?._id) ? true : false;
-}
-
-async function deleteMessage(chatId: string, messageId: string, fromId: string, forAll = false, attachments: string[]) {
-  let params: any = [
-    {
-      _id: new ObjectId(chatId),
-      "messages.id": new ObjectId(messageId),
-    },
-    {
-      $addToSet: {
-        "messages.$.deletedFor": new ObjectId(fromId),
-      },
-    },
-  ];
-  if (forAll) {
-    params = [
-      {
-        _id: new ObjectId(chatId),
-      },
-      {
-        $pull: {
-          messages: { id: new ObjectId(messageId), sender_id: fromId },
-        },
-      },
-    ];
-    if (attachments) {
-      await removeAsset(attachments);
-    }
-  }
-  return chatsCollection.updateOne(params[0], params[1]);
-}
-
-async function updateMessage(chatId: string, messageId: string, update: MessageUpdate, fromId: string) {
-  return chatsCollection.updateOne(
-    {
-      _id: new ObjectId(chatId),
-    },
-    {
-      $set: {
-        "messages.$[message].text": update.text,
-        "messages.$[message].edited": true,
-      },
-    },
-    {
-      arrayFilters: [
-        {
-          "message.id": new ObjectId(messageId),
-          "message.sender_id": fromId,
-        },
-      ],
-    },
-  );
 }
 
 async function getBlockedUsers(userId: string) {
@@ -635,50 +423,7 @@ async function blockUser(userId: string, blockId: string) {
     },
   );
 }
-
-async function unblockUser(userId: string, blockedId: string) {
-  return usersCollection.updateOne(
-    {
-      _id: new ObjectId(userId),
-    },
-    {
-      $pull: {
-        blocked_users: new ObjectId(blockedId),
-      },
-    },
-  );
-}
-
-async function provideSignedURL(
-  chat_id: string,
-  filesInfo: {
-    name: string;
-    size: number;
-  }[],
-) {
-  const filesWithSignedURL: {
-    signed_url: string;
-    key: string;
-    file_name: string;
-  }[] = [];
-  for (const file of filesInfo) {
-    const fileID = new ObjectId();
-    const fileExtension = file.name.match(/\.[0-9a-z]+$/i)?.[0] || "";
-    const path = `chat_${chat_id}`;
-    const key = `cc_${fileID}${fileExtension}`;
-    const url = await getPostSignedURL(`${path}/${key}`, "attachment", 10);
-    filesWithSignedURL.push({
-      signed_url: url,
-      key: `${path}/${key}`,
-      file_name: file.name,
-    });
-  }
-  return filesWithSignedURL;
-}
-
 async function deleteAccount(userId: string) {
-  await removeProfileImage(`${userId}.png`);
-
   return usersCollection.updateOne(
     {
       _id: new ObjectId(userId),
@@ -713,41 +458,41 @@ async function deleteAccount(userId: string) {
     },
   );
 }
-export {
-  //Collections
-  chatsCollection,
-  usersCollection,
-  // Functions
-  addConnection,
-  addMessage,
-  deleteMessage,
-  deleteAccount,
-  updateMessage,
-  updateSeenMessages,
-  acceptMessageRequest,
-  getProfileById,
-  getConnectionData,
-  connectionsData,
-  clearChat,
-  deleteChat,
-  verifyUser,
-  getChat,
-  isUsernameAvailable,
-  isEmailAlreadyRegistered,
-  oAuthGoogleLoginFinder,
-  updateOAuthProfile,
-  isUserRestricted,
-  createNewAccount,
-  setProfilePictureUrl,
-  findUser,
-  updateUnseenMsgCount,
-  updateStatus,
-  getBlockedUsers,
-  blockUser,
-  unblockUser,
-  getMessages,
-  updateProfile,
+async function unblockUser(userId: string, blockedId: string) {
+  return usersCollection.updateOne(
+    {
+      _id: new ObjectId(userId),
+    },
+    {
+      $pull: {
+        blocked_users: new ObjectId(blockedId),
+      },
+    },
+  );
+}
 
-  // Cloudflare
-  provideSignedURL,
-};
+
+export {
+    getProfileById,
+    getConnectionData,
+    verifyUser,
+    getConnectionsData,
+    isUsernameAvailable,
+    isEmailAlreadyRegistered,
+    oAuthGoogleLoginFinder,
+    updateOAuthProfile,
+    findUser,
+    createNewAccount,
+    setProfilePictureUrl,
+    addConnection,
+    updateUnseenMsgCount,
+    updateProfile,
+    clearChatMessageCount,
+    deleteChatConnections,
+    updateStatus,
+    isUserRestricted,
+    getBlockedUsers,
+    blockUser,
+    deleteAccount,
+    unblockUser,
+}
